feat: 예약 조회 결과가 없는 경우 Custom Exception을 반환한다

+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,6 +1,7 @@
 package roomescape.service;
 
 import java.util.List;
+import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Reservation;
 import roomescape.dto.ReservationRequest;
@@ -16,7 +17,15 @@ public class ReservationService {
 
     public Reservation saveReservation(final ReservationRequest reservationRequest) {
         long reservationId = reservationDao.save(reservationRequest);
-        return reservationDao.get(reservationId);
+        return getReservation(reservationId);
+    }
+
+    private Reservation getReservation(final long id) {
+        Optional<Reservation> reservation = reservationDao.findById(id);
+        if (reservation.isEmpty()) {
+            throw new IllegalArgumentException("[ERROR] 예약 조회 결과가 없습니다.");
+        }
+        return reservation.get();
     }
 
     public List<Reservation> getAllReservations() {
@@ -24,6 +33,10 @@ public class ReservationService {
     }
 
     public void deleteReservation(final long id) {
+        Optional<Reservation> reservation = reservationDao.findById(id);
+        if (reservation.isEmpty()) {
+            throw new IllegalArgumentException("[ERROR] 삭제할 예약 데이터가 없습니다.");
+        }
         reservationDao.delete(id);
     }
 }
